chore: moves e2e selectors to tests
