+++ e2e/pages/Stake/StakeView.ts
@@ -1,4 +1,4 @@
-import { StakeViewSelectors } from '../../selectors/Stake/StakeView.selectors.js';
+import { StakeViewSelectors } from '../../../tests/selectors/Stake/StakeView.selectors.js';
 import Matchers from '../../../tests/framework/Matchers';
 import Gestures from '../../../tests/framework/Gestures';
