+++ tests/selectors/Bridge/QuoteView.selectors.ts
