+++ tests/selectors/Browser/SpamFilterModal.selectors.js
