+++ tests/selectors/Onboarding/OnboardingCarousel.selectors.ts
