+++ tests/selectors/Ramps/SelectPaymentMethod.selectors.js
