+++ tests/selectors/Ramps/SelectRegion.selectors.js
