+++ tests/selectors/Settings/BackupAndSyncView.selectors.js
