+++ tests/selectors/Settings/Contacts/DeleteContactBottomSheet.selectors.js
