+++ tests/selectors/Settings/SecurityAndPrivacy/AutoLockModal.selectors.ts
