+++ tests/selectors/Settings/SecurityAndPrivacy/ChangePasswordView.selectors.ts
