+++ tests/selectors/Stake/StakeConfirmView.selectors.js
