+++ tests/selectors/Stake/StakeView.selectors.js
