+++ tests/selectors/swaps/OnBoarding.selectors.js
