+++ tests/smoke/multichain/wallet-createSession.spec.ts
@@ -7,7 +7,7 @@ import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
 import { withFixtures } from '../../framework/fixtures/FixtureHelper';
 import MultichainTestDApp from '../../../e2e/pages/Browser/MultichainTestDApp';
 import { SmokeMultiChainAPI } from '../../../e2e/tags';
-import MultichainUtilities from '../../../e2e/utils/MultichainUtilities';
+import MultichainUtilities from '../../helpers/multichain/MultichainUtilities';
 
 describe(SmokeMultiChainAPI('wallet_createSession'), () => {
   it('should create a session with Ethereum mainnet scope', async () => {
