+++ tests/smoke/multichain/wallet-invokeMethod.failing.ts
@@ -23,9 +23,9 @@ import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
 import { withFixtures } from '../../framework/fixtures/FixtureHelper';
 import MultichainTestDApp from '../../../e2e/pages/Browser/MultichainTestDApp';
 import { BrowserViewSelectorsIDs } from '../../../app/components/Views/BrowserTab/BrowserView.testIds';
-import MultichainUtilities from '../../../e2e/utils/MultichainUtilities';
+import MultichainUtilities from '../../helpers/multichain/MultichainUtilities';
 import Assertions from '../../framework/Assertions';
-import { MULTICHAIN_TEST_TIMEOUTS } from '../../../e2e/selectors/Browser/MultichainTestDapp.selectors';
+import { MULTICHAIN_TEST_TIMEOUTS } from '../../selectors/Browser/MultichainTestDapp.selectors';
 import { waitFor } from 'detox';
 import FooterActions from '../../../e2e/pages/Browser/Confirmations/FooterActions';
 import { isHexString } from '@metamask/utils';
