+++ tests/smoke/multichain/wallet-notify.failing.ts
@@ -31,7 +31,7 @@ import { SmokeMultiChainAPI } from '../../../e2e/tags';
 import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
 import { withFixtures } from '../../framework/fixtures/FixtureHelper';
 import MultichainTestDApp from '../../../e2e/pages/Browser/MultichainTestDApp';
-import MultichainUtilities from '../../../e2e/utils/MultichainUtilities';
+import MultichainUtilities from '../../helpers/multichain/MultichainUtilities';
 import Assertions from '../../framework/Assertions';
 import { DappVariants } from '../../framework/Constants';
