+++ wdio/screen-objects/SendSolanaScreen.js
@@ -2,7 +2,7 @@ import AppwrightSelectors from '../../tests/framework/AppwrightSelectors';
 import AppwrightGestures from '../../tests/framework/AppwrightGestures';
 import { expect as appwrightExpect } from 'appwright';
 import TimerHelper from '../../appwright/utils/TimersHelper.js';
-import { SendActionViewSelectorsIDs } from '../../e2e/selectors/SendFlow/SendActionView.selectors';
+import { SendActionViewSelectorsIDs } from '../../tests/selectors/SendFlow/SendActionView.selectors';
 
 class SendSolanaScreen {
